writing mutation and split the buttons to their own js file
